Fixes for manta tests.
Tests were not doing image comparisons and they could not be made to fail.
The image problems were due to various path issues (and a lack of
checked in images in the DATA_ROOT baseline.) The fact that they can not
fail is deeper and probably is caused by Manta hickacking exit codes
An alternate workaround is to print and grep for error messages via CTest.

I also converted the standalone executables to a combined test executable
to cut down on solution clutter.

Change-Id: I725d53c8888b779809c0f1c35e5d4deeb4229e62