Fix server side rendering for manta.
This reverts commit 5cdabd4941c9 which I believe was a merge mistake that happened
when EDL and Manta were both updated to the new views and representations architecture.
The code mistake is that the server side can not use the client server compositer to
do the server side compositing. Also, Manta does need to use the old compositer
classes since the IceT ones assume GL*.

*That statement isn't quite true anymore but making vtkManta use IceT is too much
effort at the moment.

Change-Id: I3faac826d6835f74a10cf0bfab0a5be9aa12db73