Fix streaming regression after the collaboration merge.
Client server mode was not working, for either server side or client side rendering,
after collaboration. I used to instantiate the streamed rendering management filters
only on the server by only sending the stream that instantiated them to the server.
This commit does the corresponding check within the new architecture.

Change-Id: Ic0c48969e6550346c84ca9aa05b6d55f15eed3cf