Fixed BUG #12105. UpdatePipelineInformation() was being called incorrectly.
UpdatePipelineInformation() is implemented only on vtkSMSourceProxy. However on
the SI side, I was being lazy and implemented that on the vtkSIProxy instead of
vtkSISourceProxy. Consequently, UpdatePipelineInformation() would get called on
the subproxies that were not SourceProxies on the client side. This was causing
the errors when loading states. Fixed that.