Connect pipelineBrowserWidget earlier to pqServer
This allow a valid collaborative initialisation, otherwise
when additional source were created inside the ProxyManager
they couldn't be inserted in the pipelineBrowserWidget
because the server was not registered yet.

As the server is registered earlier, its name will be invalid
and will trigger a pqPipelineModel->updateData(NULL) which
I've updated to make sure every item gets updated in that
particular case.