fixes for streaming tests
Window captures for streaming did not know to iterate, so results
results were wrong and not reproducible. This change fixes that.

Also the test restricts the number of passes to validate piece ordering.
However the absolute number of pieces is a function of the number of
processors, which is different on different dashboard machines.
So I prevent that test from running in parallel and add new tests that
show all the pieces in order to exercise streaming in parallel.

Change-Id: I5b39b5893973ba1542ce1ffb3666f115b4453594