Corrected a nasty bug when calling processEvents.
The reaction was connected to the sources modified signal rather than data loaded signal. This previously was only working because the code flow didn't go to the Qt event loop before the source finished loading.
Now with progress calling processEvents the reaction was attempting to create filters before the source had finished loading and would fail.