UpdateDataEvent is now sent after all servers are done processing.
We don't want the UpdateDataEvent sent while servers are still operating on the respective vtkObject. Because we presume that UpdateDataEvent is sent when the object is valid and we can query it for information. This bug has been hidden for along time since we never got back into the Qt Event Loop while the server was processing before.