Fixed deadlock with 2 clients in local-render mode.
There was deadlock with 2 clients in local render mode and after interacting
with the first client, the user tried to interact with the 2nd client.  The
issue was that vtkUnstructuredDataDeliveryFilter was incorrectly recording that
LOD data was delivered to 2nd client in ProcessViewRequest. ProcessViewRequest
gets called on both delivery filters in
vtkPVGeometryRepresentation::ProcessViewRequest and hence LODDelivery filter was
assuming it had delivered to the 2nd client prematurely.

Change-Id: I77d53118b13787fc1b1c786b2c49053a15e5290b