Fixed BUG #12206. Fixed freeze when property was changed from python.
When the user changed a property from Python shell and started interacting,
ParaView would hang when user started interacting in client-server mode. This
was happening because the interactor styles would call mapper->update() on
interaction. This caused the client side delivery filters to execute
prematurely. Fixed that problem by doing two things:
1) Put an update suppressor before the mapper to avoid such stray updates for
   causing havoc.
2) vtkPVDataRepresentation subclasses now mark their delivery filters dirty in
   RequestData and not MarkModified(), since the real input data is not
   available until RequestData().

Cleaned up vtkPVUpdateSuppressor API to remove all the request related ivars. We
don't use that anymore. UpdateSuppressor now simply becomes an update-supressor
not a update-conditioner as it was before.