Making ProxyDefinitionManager a Remote Object.
Renamed vtkPVProxyDefinitionManager as vtkSIProxyDefinitionManager.

ProxyDefinitionManager was not a true remote-object. Consequently there were
scatterred hacks to ensure its state is kept up-to-date among processes.
Cleaning up that code to ensure that the ProxyDefinitionManager uses the
RemoteObject API for that.

Also ensuring that when a plugin is loaded, the XML definitions are synchronized
between client and server.