Fixed BUG #12238. Load plugins before server-added is noted.
Fixed pqServer initialization sequence to ensure that the 'resource' information
for the pqServer is set correctly when serverAdded() and other such signals are
fired by the pqServerManagerModel.

Fixed pqPluginManager to load plugins in preServerAdded signal handler to load
plugins before the application realizes that a new server connection has been
established.