Fixed BUG #12241. Multiple connects was causing issues.
The problem was that the queued-connection was causing the newly connected
server to be disconnected as well. In general this server-connection code is
very hard to debug. The complication exists due to the fact that when we
originally wrote this code, we didn't allow modal dialogs in ParaView. That
resulted in very hard to read code and bugs keep on creeping in. We need to bite
the bullet and clean it up at some point.