WIP: Added mechanism to disable widgets/actions in "slave" mode.
We needed a mechanism to disable widgets and hide actions when the processes is
not a master in collaborative mode. This commit adds support for that. The
developers need to set the keys PV_MUST_BE_MASTER_TO_SHOW and/or
PV_MUST_BE_MASTER as dynamic properties of the widgets and actions that need to
be available only when the application is the master.

Haven't updated all GUI components to specify these keys yet, but the mechanism
is in place.

Change-Id: Ia8602d2e21c770ac4360ebd74132fe2009d9d644