Making unstructured gradient filter available to all data sets.
The unstructured gradient filter can handle all types of data sets
but it is slower than the image data gradient filter.  It has
more functionality than the image data gradient filter so I'm
changing ParaView's access to it such that it can be used
with vtkDataSets instead of just vtkPointSets.