Making delivery request client-driven.
Now, representations should not deliver any data to the "rendering nodes"
unless the driver tells it to. Add a new DELIVERY phase to the renderview
rendering phases. The DELIVERY phase is called on only those representations
that the driver thinks need to delivery. Also any representation is required to
do delivery in this phase (if it does indeed do delivery).

Updating most render-view representations to use DELIVERY() stage for
data-delivery.