Modify delivery filters when changing subdivision level.
When the nonlinear subdivision level is changed in the display panel,
that propogates to an internal geometry filter that extracts the
surface and wireframe.  This changes the geometry that is about to be
rendered, which means in client/server mode with geometry delivery, the
data will have to be resent to the client.  The server knows about this,
but the client does not.  To get around the problem, simply mark the
delivery filters as modified to force both client and server to transfer
geometry.

Change-Id: Iec664665f2aa2e117d4bb0145355a3025f0fe5fc