Avoid spreadsheet view updates until ready.
pqQueryDialog::runQuery would call StillRender, however the resulting progress
events would invariably cause the tree view to repaint which could start
querying the spreadsheet model while it was processing the StillRender(). This
was causing random dashboard failures. Fixed that by ensuring that the model on
the view is not set until we have done the StillRender()