BUG: Change the export VRML test so that it matches the baseline.
Previous versions of ParaView actually had a bug in the loading of vrml files that had an active color. The first render would not be correct. Now that we have proper progress and more calls to process events the proper color started to show up. Instead of updating the baselines with new images and breaking master and next instead I am changing the test