Fix UndoRedo test 2
The test was failing during the Undo of pqCloseViewUndoElement
because the associted representation of the view were not restored
as part of the view since the ProxyProperty were not allowed to
Locate those proxy in order to prevent the creation of them.

To fix it, we disable that limitation during the Undo call only.

The creation prevention is related to some widget issue in multi-client mode.
But I'm not sure if this is still valid as we disable widget for slaves !