Fix UndoRedo3 by creating a context based proxy locator
For each undoset Redo/Undo action we create a custom proxy locator
that will use in priority the state that are present in that undoset.

Therefore if we load a state of a proxy that require another one that
has a given state inside the undoset we will have the correct state.

Otherwise we were using the last state that was used in the session
which was basically wrong in some case.