Fixed BUG #11347. Fixes domain issues for statistics filters.
vtkSMArrayListDomain was looking at value properties but was not doing
enum-conversion correctly. Fixed that by adding a static method for conversion
to vtkSMInputArrayDomain.

vtkSMFieldDataDomain now updates the unchecked value as well as checked value.
This fixes the issue with the list being empty if dataset only had cell-data. In
general, we need to fix the unchecked-mechanism, this is just work around.