Fix 12462 Selection in Spreadsheet
The issue was related to some proxy that was getting deleted by
a slave inbetween two ClearInputs() SetInput() calls.
To solved that we use Reference count on the server side
to prevent object to be deleted if one client is still using it.
So now, each time a new client connect to the server, it increase
the reference count of the SIObjects that it is using.
In the case of a client crash a automatic garbage collection will
occurs and will decrease the reference count of the objects that
was still held by the client before the crash.