Fixed BUG #12491. Making CacheKey and UseCache non-internal.
This fixes issus in mutli-client mode when using cache. Also, there was no real
need for these properties to be internal since their state is always returned to
the default value by the animation playing code, so we don't have to worry about
enabled cache being saved in state files.

Change-Id: I38273fecb9d6c519e2a3f75947dbdd45a9812502