Fixed BUG #12546. RequestData was called twice in parallel.
The issue was introduced by fix for BUG #11811. We were now requesting
ghost-cells in representations to use the ghost-cell information when available
during rendering. However that meant that all the request changed since previous
update on the pipeline, since the GUI would call UpdatePipeline() when creating
the representation but without the extra ghost cells.

Fixed this issue by making the GUI also ask for ghost-cells in a similar fashion
to vtkGeometryRepresentation. This was done by updating
vtkSISourceProxy::UpdatePipeline to request ghost cells using the logic in
vtkGeometryRepresentation.

vtkGeometryRepresentation was checking for data-type when deciding whether to
ask for ghost cells. That meant that vtkSISourceProxy couldn't use it since the
algorithm had not produced any data yet. Anyways, the datatype check was
unnecessary since we were using the presence of WHOLE_EXTENT() key to exclude
any structure data pipelines. So removed that type check. I verified that it
doesn't lead to any wrong ghost cell requests when contouring, for example.