Updating in-app user's guide to 3.12.
The wikibook to openoffice export is buggy such that the improved
wiki content's images are missing in the converted help doc.
So, I've stripped down the in-app doc to the essential contents,
manually fixed those, and refered the reader to the full version
that they can get online or with our packages.

I've also updated what remains to reflect 3.12 and to correct other
longstanding formatting problems.

Change-Id: I8a86cec8f92c5ed965d6f57c79ae1b3180b499f1