FIxed BUG #12638. "random" default wasn't working correctly.
The code lost the fact that "random" was chosen for default in the config. Fixed
that. This chunk of code seemed to update teh pvsc file itself when values
changed. Not sure why that was done. We should only save user choices in the
registry, not in the pvsc itself!!