CoProcessing: Looking through the filters to determine available writers.
Removed Writers.xml since all of that info was already in servermanager.xml.
Now I go through all proxy definitions and look at the Hints to see
if there is a CoProcessing hint.  If there is I assume that this is
a stub for a real paraview writer that can be used during coprocessing.
This also has the advantage that writer plugins should automatically
work with coprocessing as long as they have this special stub.
Look at Plugins/CoProcessingScriptGenerator/Resources/servermanager.xml
to see how to do this.