CoProcessing: Fixing problem with stub only definitions of CP writers.
A user can create an XML only plugin that describes the writer.
This results in a stub of the real proxy which doesn't have to
exist in the build used to create the coprocessing script but
must exist in the build or a plugin that is loaded while
running the simulation code.

Change-Id: I2dd5a90106d67270332201251564ec8d16a72a0e