Fixed ability to create views on different sessions at the same time.
vtkPVView now creates a vtkPVSynchronizedRenderWindows instance per session.
This ensures that multiple views exisiting on different sessions can be created
simultaneously.

Fixed vtkPVSynchronizedRenderWindows to set the name for each VTK-window. Now
all windows show up with nicer names that make it possible to determine which
window corresponds to which process.