ENH: Change ParaView to use the Smart Volume Mapper in VTK
This will allow ParaView to use the smart volume mapper inplace
of both the fixed point and the GPU mappers.  Both of these
mappers are contained inside of the smart mapper as well as a
Texture Mapping based approach.  This change also simplified
the display panel since some of the properties are determined
by the smart mapper.

Change-Id: I10b3932635fc0f2f24a0d5e4deabf4e7054a2ce5