Fix for vtkWindBladeReader bad memory access and compiler warning.
Got rid of potential bad memory access and unused variable compiler warning.
It didn't happen with the test but there was a possibility that
the reader would try to dereference a null pointer.  Also, getting
rid of an unused variable compiler warning and moving local variables to
more limited scopes.

Change-Id: I52e047d37b6740253a1c883afa9a5bde75411e80