Fixed BUG #12720. Minimize reads on satellites.
All processes were reading first file to gather meta-data. This caused issues
when running in parallel on large number of cores. Fixed by reading the file on
root node and then broadcasting the gathered information to all nodes.

Structured the code slightly to avoid processing of the meta-data when timesteps
changed.