FIX: ParaView compiles with new VTK API.
1. DataObject->SetWholeExtent
This method is removed as WholeExtent refers to the pipeline and dataObject does not hold any reference to pipeline

2. DataObject->GetInformation
Use this method when you want to change any information pertaining to the dataObject.
For pipeline-related changes, get the information-object from the algorithm that produces the dataObject.

3. DataObject->GetProducerPort
This method is removed as the dataObject does not hold any references to pipeline information.
Instead, if a pipeline connection is needed, one can use the information key PRODUCER in vtkExecutive as follows:

vtkExecutive::PRODUCER()->Get(pipelineInfo, producer, producerPort());

If the dataObject does not have a producer, a vtkTrivialProducer can be instantiated and its the dataObject should be set as its output.

4. DataObject->SetPipelineInformation(info)
This can be done by

info->Set(vtkDataObject::DATA_OBJECT(), DataObject)

FIX: Added a data structure to maintain shallow copies of input data to avoid memory leaks in vtkPVDataRepresentation.
It is a map from (port index, connection index) to (original input data port, shallow copy port).

FIX: Making a shallow copy of input in vtkDataRepresentation was causing us to lose references to rendered objects.
Directly connected the pipeline without making a shallow copy.

Added vtkAlgorithm as an argument to the signature of ExecuteBlock in vtkPVGeometryFilter to use the input algorithm directly.

Conflicts:

	ParaViewCore/ClientServerCore/vtkMPIMoveData.cxx
	Qt/Components/pqViewManager.cxx
	Utilities/Xdmf2
	VTK

Change-Id: I4fce8cb700e5ac9a5afaa3423eab9d3976fdb291