Fixed problem of trying to autoload coprocessor plugin.
Problem was that the coprocessor plugin was trying to modify a currently
non-existent main window to add menus to it if it was set to autoload.
I added a timer if the main window wasn't available yet so that the
plugin wouldn't be loaded until the ParaView GUI was in the proper
state.

Change-Id: I57f726d7015b43d5818fd26d3dba2d39b0f38e20