Fixed BUG #12773. Comparative View was not updating correctly.
When the cue was changed, we were not calling MarkModified() on the cue proxy.
Consequently, the view never thought it was dirty and needed to update and hence
never did. Fixed that by exposing API from vtkSMComparativeAnimationCueProxy to
chaneg values and ensuring those methods call MarkModified(). Hidding the access
to internal vtkPVComparativeAnimationCue so that we don't do similar mistakes
again.