Combine field selection into one property
This combines the field selection property into a
single QStringList property with two items instead
of using two separate QString's for storing the attribute
mode (e.g. Point Data, Cell Data) and attribute scalars.

This fixes a bug in which an infinite loop was encountered
when changing the scalars between the summary and properties
panels.

Change-Id: I8316c58cb47995f08edd0a383e8180536352a521