Fix for rendering problem in batch, parallel, symmetric mode.
If paraview is being run in batch, parallel, symmetric mode it is
possible that some processes may have 0 points and in this case
there probably won't be any point or cell data available to
color by.  If the user tries to set the ColorArrayName property
on these processes it would crash.  This change fixes that problem.

Change-Id: I58886bbb6f02c4dfd9ca7ecf5c6b05bd824d0be0