Static plugins not detected correctly on client.
Plugins that were linked in statically were not being detected on the client
side. Consequently there was a plugin mismatch between client and server. Fixed
that by refreshing the list of plugins on the client side every time a session
was created.

This uncovered a bug in vtkPVPluginsInformation::Update() with statically linked
plugins that have their filename set to "linked-in". Fixed that as well.

Change-Id: I5d1118fb13e934b3c0d8e462709465ff033ee7ca