Added support for comparing images/widgets in XML testing.
Python test playback is tedious since it uses multiple threads. Most of the
tests in Python for the GUI are done so that we can compare images or multiple
view widgets. Adding API to pqXMLEventSource so that it's now possible to do the
same using XML. New XML nodes "pqcompareview" and "pqcompareimage" can be used
for that purpose.