Ensuring that state is pushed before firing events.
This is necessary in collaboration mode since other we are deliberately
conveying state changes to the other clients out of order (since any handler
code for the registeration event, for example, will push its state changes,
before the proxy registration state change message is pushed.

This causes issues with tabbed views in multi-clients mode when splitting and
creating new views.