Remove unregistered views.
When a view proxy was unregistered, pqMultiViewWidget wasn't reclaiming the
frame. Although (thanks to weak pointers) this wouldn't cause any major havoc,
the frame ended up without any "Create View" buttons. Fixed that.

The problem was evident when one split a view, create a 3D view and then undid
the creation of the 3D view.