Fixed issues with charts not rendering anything on creation.
The problem was that the axes range for the charts wasn't set correctly. This
was because the chart tried to reset its axes during creation based on the
widget's size. However the widget didn't have any valid size (when the layout
had changed) since the layout update events weren't handled by Qt yet. So we
ensure that pqMultiViewWidget processes all pending Qt events when the layout
is updated.

Change-Id: I362da80d6d5707550673440c6c8489a602a7e505