BUG #12803. Added support to compute average material density.
Note that the average density is not computed by averaging the material
densities, but by summing the mass of each material in a cell and dividing by
the cell volume. If void is present, the volume of the void has to be subtracted
from the cell volume before computing the density.