Fixes segfault with chart views are created by default in client-server mode.
In client-server mode, ParaView was segfault with a chart view was the first
view (or only view) created. This was because vtkPVSynchronizedRenderWindows was
setup in an "enabled" mode that would cause communication messages to be sent
before the vtkPVContextView had an opportunity to disable the
vtkPVSynchronizedRenderWindows in it's Render() method. Fixed that.