Fixed errors reported when using command startup for reverse connect.
The problem was due to the fact that the client launched the server process
before opening the socket to accept the server connect request. This caused in
error messages from  the server, more often than not. Although the server would
eventually connnect, it was quite misleading. Fixed that. Now pqServerLauncher
opens up  the port before launching the server in reverse connect mode.