Use real hostname instead of "localhost" by default.
Instead of using "localhost" as the default server/client host name, the code
now uses hostname. Consequently, one can now simply copy the URL printed out the
server and paste it  as the -url argument to the client to connect to it (so
long as the client can connect to the server node).