BUG #12933. Selection on PlotOverLine filter doesn't work.
The issue was that the PlotOverLine filter is vtkSMCompoundSourceProxy and the
selection code was ending up selecting the internal proxy. This was due to the
fact that vtkSMOutputPort was not being told the vtkSMCompoundSourceProxy was
it's "owner". Fixed that by add new API on vtkSMOutputPort that's used by
vtkSMCompoundSourceProxy to indicate that the vtkSMCompoundSourceProxy is the
"source" of the output port for all the rest of the world.