Fix layout in multi-client + mutli-server mode.
There were cases when the view was created before the layout. That caused havoc
in multi-clients mode since the other client ended up creating a separate layout
for the same view thus ending up with two layouts.

This fixes that issue be ensuring that we always create a new layout, if none
exists before creating a view.