Fixed issues with test playback in collaborative mode.
The problem was that the participant ended up assigning a new frame to the view
as the collaboration state was being loaded, thus changing the layout in
unintended ways. Fixed that by ensuring that a frame is not assigned to a newly
registered view if it's proxy has already been assigned to a layout.