Refactoring pqPropertyLinks.
Added a new pqPropertyLinks2 class (which will eventually be renamed to
pqPropertyLinks) that acts as a replacement for pqPropertyLinks. The new design
uses a simplified logic (and reduced code) and does not rely on timers or
delayed connections of any sort.

Needs unit-tests and exhaustive testing before becoming the replacement for
pqPropertyLinks.