Adding a test that goes through the full coprocessing workflow.
The test creates a script with the plugin.  This test script outputs
both a pvtp file and a screenshot (png).  Then it runs a driver
code to simulate coprocessing using the generated script.  Finally,
it compares the image output to a baseline as well as checks the
pvtp file for the proper amount of points and cells in the grid.

Change-Id: I252ac9c0db7e9dc1a47f6765c0f7af33d3e4ae8b