Charts now update on tiles when changed.
Plots on the tiles were not updating. This is because we were not providing
valid data to those on the server side. Fixed to clone data on server nodes when
in tile-display mode.

Added a new key that the vtkPVContextView uses to communicate to its
representations that we are rendering on a tile display and the data must be
cloned. This is only done for tile displays since in other cases it's not
necessary.