Refactoring axis range properties for charts.
Charts did not save the updated axes ranges when the user interacted with the
charts in the state files. This was because the property for axes range was
not updated on interaction. Changing the entire infrastructure to get this
working.

Hence forth, as soon as the user interacts with the charts, the axes become
"fixed" and don't auto-update unless user hits "reset-camera".

There are 4 new properies LeftAxisRange, RightAxisRange, TopAxisRange, and
BottomAxisRange that can be set to empty to enabled "auto" ranges or set to
specific values to use those ranges.

As soon as user interacts with the view, vtkSMContextViewProxy listens to the
interaction event and updates the axes range properties using the current range,
thus "fixing" them to the chosen values.

Also creating special vtkContextViewInteractorStyle subclass to ensure that we
call StillRender() on proxy every time the user interacts with the view.

All these fixes now ensure that charts on tile displays show current axes ranges
and update as user interacts with the charts.

WIP: This breaks the synchronization of chart viewports in collaboration mode.
That needs to be fixed.