Fix management of animation scene creation
Animation scene creation was triggered in case of multi-server
when we were disconnecting to one of our remote server.
The first attempt to fix that issue was incorrect.

Now we check if the server/session is in a state that illustrate
that is going to be removed. If so, we simply prevent that scene
creation to happen.

Change-Id: I7e15f485ef9c8d6c1fd7ed972962c64f5e98b075