Brought back the color table for single scalar.
When a dataset has a single scalar value, a color-button was used, which
seems to cause confusion to some users since they used to see a table with
single color for that. This commit bring back the old behavior, and the
user can click on the table to change its color.

Change-Id: I0e6bcd6891d75b508ac2e169871c6f841c74fed1