Refactor vtkCompositeRepresentation.
vtkCompositeRepresentation forwarded calls to the active representation. This
kept the internal representation hidden from the view. That makes it complicated
for delivery mechanism to handle it correcly. All this was done for selection to
be able to detect the selected representation proxy. So fixed the selection code
to deal with composite representation correctly and vtkCompositeRepresentation
no longer forwards calls to internal representations, instead internal
representations are added to the view directly.