Add new properties panel dock
This adds a new dock to the main ParaView window which
contains the new properties panel widget. This new panel
replaces the old properties and display panels.

The new properties panel is still a work in progess and
is not enabled by default. To use the new properties panel
pass the '--use-new-panels' argument on the command line.

Change-Id: I977d7c5a598247b95664e20ab30d30f82b82013d