Modified some default mouse button actions for chart selection
By default in chart selection, the right-mouse-button will do selection
for rectangle or polygon selection in charts. However when one of the
Subtraction, Addition, Toggle selection mode is used, the
left-mouse-button will be used for selection instead. Also, one of the
right-button selection actions will always be active in charts.

Change-Id: I4e64d4c2c122ce25efe58c9c560cd214c3436d65