Replace use of find_file(...) with macro to cope with file moves
find_file(...) sets a cache entry onces a file is found. If a files
moves for example in VTK this causes problems. So instead walk
through the ${module}_INCLUDE_DIRS check if the file exists. This
can cope with moves at the value of ${module}_INCLUDE_DIRS will be
updated.

Change-Id: Ifb83c87e3f68a7339f4a800d395f6cffb0f6922e