Fixed Slice representation in client-server modes.
vtkPVDataDeliveryManager was ending up not requesting new data when slice
parameters were changed. This happened because the mtime we used in
vtkPVDataDeliveryManager used the data-time, which clearly didn't change on the
client after every update. Since vtkPVDataDeliveryManager::SetPiece() uses the
executive's data-time to check if data was really updated, we shouldn't use the
data mtime in vtkPVDataDeliveryManager's SetDataObject() call.