Fixed issue with comparative views.
After the latest round of refactors, the comparative views were broken. This
happeneded since we removed the code that forwarded the cache related values to
the internal representations in vtkCompositeRepresentation. That was a mistake.
Consequently, when the cache keys were set on the representation's properties by
the vtkPVComparativeView, they pretty much had no effect and the comparative
view ended up rendering incorrectly. Fixed that.

Change-Id: Ib89fdb06b1fd4438a2c19cd3a567b862051157a0