Refactor vtkSMProxy::SetupExposedProperties()
This refactors the SetupExposedProperties() method in the
vtkSMProxy class and moves most of the proxy setup logic
to a new method called SetupExposedProperty(). This allows
for property groups to set up each of their properties
individually.

This also fixes a bug when a subproxy had exposed properties
with multiple property groups.

Change-Id: I633e29d0e868d86821414726a4576ba4042e751e