BUG #12542. Added support for interactive mode using outlines.
Removing ClientOutlineThreshold - it was used to control when to deliver
outlines to client for interactive renders. Instead, added a checkbox to the
"LOD Parameters" dialog to indicate that when using LOD, use outlines. When
checked, if the view decides to use LOD rendering for interactive renders (based
on the LOD threshold), it will use outline for the LOD instead of
quadric-clustering.