Add new data fields for custom cube axis behavior
Those data annotations are used to provide an oriented bounding box
which will allow the cube axis to properly fit the object bounds without
relying on the orthogonal axis.

To illustrate the usage of those annotation, we've created a simple plugin
that define a custom source that could define arbitrary base axis.

Here is the list of the available annotation:
 - 3 vectors use to define the base
 - oriented boonded box (using the defined base)
 - Axis title
 - Axis center (Still need to implement that behavior on the cube axes)

Change-Id: I93f4f5f73ca5106ed7c2557941dfe4e4515e94fc