Fixes for the coprocessing.
Resetting the python trace after the coprocessing test is created
since that was causing problems on exiting from paraview as well
as test failures. Also a change for the xml script since the Qt
object names have changed.
Changed the parallel script test to run in serial (i.e. to
run when no other tests are running, not running with a single process).
They also use the same script now and so I can delete PythonScriptTest2.py
since that only output a different file name.

Change-Id: I6bea171bbd8e2cf0f74d33ea77d6ab81b29bbced