BUG #13412: Fixing dependency issues when building documentation.
Resolving build problems with Ninja when bulding embedded documentation for
ParaView. The issue was deeper than just a Ninja problem.

Docmentation/CMakeLists.txt was generaing a paraview.qhp file and a resource
file under Applications/ParaView was dependending on it. That's a tricky
dependency since Applications/ParaView doesn't see the rule to generate
paraview.qhp. We overcome the issue by adding a phony ParaViewOnlineHelp target
that we made paraview depend on and than ensured that the qhp was available by
the time Applications/ParaView's resource file looked for it.

Fixing the issue by moving Documentation to Applications/ParaView. The
documentation is indeed for the ParaView application hence makes sense its
included in the application's sub-dir. We build a static library containing the
documnetation resource that's initied by ParaViewMainWindow constructor.