BUG #11065. Added support for "enhanced outline" for AMR.
For AMR, instead of showing merely the outer outline, we now show the outer
outline etched with the dataset outlines. Unlike in past, the internal block
outlines are not shown, merely those on the outer surface.

Also for overlapping AMR, outline doesnt' need any heavy data. It merely uses
the AMR box information.