Enabling passing scalars for unstructured grids in the contour filter.
Before the contour filter for mixed multiblocks (some structured grids
some vtkUnstructuredGrids) wouldn't allow you to pass the scalars
when contouring and the output would be odd in that the structured
grids would have scalar values but the unstructured grid wouldn't.
That has been fixed in VTK and this now exposes the extra options
in the gui.

Change-Id: I6624bbf27b8b64a953a0af8c08107135713636d7