Generate faces for outlines of AMR dataset.
For BUG #0011065, we need to enable culling of front faces. For that its
essential that the outlines have front/back faces to begin with. Hence, making
vtkPVGeometryFilter generate faces for outlines for AMR datasets. Since
vtkOutlineRepresentation is anyways set to show "wireframe", the resulting
visualization remains unchanged.

This can be theoretically done for all data types, but leaving that for a later
date.