BUG #13451. Fixing modules needed by ParaView.
Also fixing the logic to enable modules in VTK. Now we no longer force the cache
variables. That approach was bad since once forced to on, those modules stayed
enabled and user had to manually disable them. Instead, now we rely on
COMPILE_DEPENDS to ensure that the necessary vtk modules are enabled.