Adding support for streaming in client/server configurations.
"AMR Blocks" streaming now works in client-server with local or remote
rendering. This updates the logic for delivering streamed pieces to rendering
nodes. Although both remote/local rendering are supported for streaming, if the
rendering mode is changed in between, the streaming is not restarted (currently)
and hence the blocks already delivered will not be available on the other
processes.

Tile display configuration also works.