Cleaning up logic to update enable state for filters menu.
Adding new signal on pqActiveObjects that makes it easier to detect with the
active source updates its data. pqFiltersMenuReaction now uses that signal to
decide when to update the enable state of the menu.

This fixes the issue with some Windows machines (amber12, release, 32bit) with
updating the enable state of the menu for certain tests
(pv.SpyPlotHistoryReader.StreamTracerUpdates).