Revert "Fixed issues with charts not rendering anything on creation."
This reverts commit ec0a74d560eadeda522f6bf604aa2c02c4a15a8a.
The original commit was added to fix rendering of charts. That doesn't seem to
be an issue anymore. Calling QCoreApplications::sendPostedEvents() causes issues
on Windows since it tries to render QVTKWidgets are they are being destroyed,
for example. Reverting ec0a74d5 removes this change that causes the issue and
QVTKWidgets get a chance to die before stale events are processed.